feat(gql): ISO GQL conformance, correctness + OOM fixes, an RDF loader, and correlated-subquery decorrelation
Extends ragedb's GQL support toward ISO/IEC 39075 and hardens the executor for
the LDBC SNB workloads. The IC/IS set runs; this broadens coverage into
BI / SPB / finbench.

Highlights:
- Variable-scope subqueries: CALL (imports) { ... } and
  CALL algo.propagate(...) YIELD node, value, depth (a value-propagating BFS).
- Temporal component accessors, duration(), date(); a list toolkit
  (comprehensions, all/any/none/single, range(), size(list), list[i]).
- CAST, IS LABELED, OFFSET paging, FOR expansion + list literals, EXCEPT,
  explicit GROUP BY, edge-orientation predicates, a scalar-function library.
- Weighted ANY SHORTEST ... COST (Dijkstra); path mode after the path variable.
- An N-Triples / N-Quads RDF loader (Lua LoadNTriples).
- OOM-safety fixes for the LDBC IC crash class (bounded shortest-path frontier,
  streamed var-length aggregation, paged label scans).
- Correlated-subquery (COUNT{}/EXISTS) precompute decorrelation -- once per
  distinct correlation key instead of once per row (BI Q12 ~2x on SF1).
- Correctness: node .key resolves via getKey(); a FOR variable binds before a
  MATCH that anchors on it; a node property converts to its concrete Lua type.
- GET /version reports the build's git SHA.

See the PR description for the full breakdown.

+++ src/gql/optimizer/DomainConstraintReasoner.cpp
@@ -373,7 +373,7 @@ std::shared_ptr<LogicNode> clone_and_rename_logic_tree(const std::shared_ptr<Log
     copy->kind = node->kind;
     if (node->kind == LogicKind::LEAF) {
         int old_id = node->literal_id;
-        const auto& old_pred = predicates[old_id - 1];
+        const auto& old_pred = predicates[static_cast<size_t>(old_id - 1)];
         std::string new_var = (old_pred.variable == c_var) ? q_var : old_pred.variable;
         copy->literal_id = get_or_create_predicate(new_var, old_pred.property, old_pred.op, old_pred.value, predicates);
     } else {
@@ -413,7 +413,7 @@ std::shared_ptr<LogicNode> build_constraint_logic(const GqlQuery& c_query, const
 
 bool update_state(PropertyState& state, BinaryOpKind op, bool assigned_val, const property_type_t& value) {
     if (std::holds_alternative<int64_t>(value) || std::holds_alternative<double>(value)) {
-        double val = std::holds_alternative<int64_t>(value) ? std::get<int64_t>(value) : std::get<double>(value);
+        double val = std::holds_alternative<int64_t>(value) ? static_cast<double>(std::get<int64_t>(value)) : std::get<double>(value);
         Interval limit;
         if (op == BinaryOpKind::LT) {
             if (assigned_val) {
@@ -490,7 +490,7 @@ bool check_theory_consistency(const std::map<int, bool>& assignment, const std::
     std::map<std::pair<std::string, std::string>, PropertyState> states;
     for (const auto& [var, assigned_val] : assignment) {
         if (var >= 1 && var <= static_cast<int>(predicates.size())) {
-            const auto& pred = predicates[var - 1];
+            const auto& pred = predicates[static_cast<size_t>(var - 1)];
             auto key = std::make_pair(pred.variable, pred.property);
             if (!update_state(states[key], pred.op, assigned_val, pred.value)) {
                 return false;
@@ -658,7 +658,7 @@ void DomainConstraintReasoner::domain_constraint_reasoning_pass(GqlQuery& query)
     
     // Convert logic tree to CNF via Tseitin
     CNFBuilder builder;
-    builder.next_var = predicates.size() + 1; // Start Tseitin vars after predicates to prevent clashes
+    builder.next_var = static_cast<int>(predicates.size() + 1); // Start Tseitin vars after predicates to prevent clashes
     int root_lit = builder.convert(root_formula);
     if (root_lit != 0) {
         builder.clauses.push_back(Clause{{root_lit}});
